make NameForm presentation component
