--- src/features/Authentication/components/Login/components/NameForm/index.ts
@@ -1,1 +1,0 @@
-export * from "./NameForm";
