--- src/features/Authentication/components/Login/index.ts
@@ -1,2 +1,0 @@
-export * from "./components";
-export * from "./Login";
