+++ src/features/Authentication/index.ts
@@ -1,5 +1,6 @@
-export * from "./Authentication";
+export * from "./Header";
 export * from "./authStore";
 export * from "./components";
 export * from "./loginSaga";
 export * from "./WithAuthentication";
+export * from "./Login";
